Fix Hyrule Warriors Rupee Address
Previously it was set to a length and size of 5 which grabbed the wrong value. As can be seen on [this gbatemp thread](https://gbatemp.net/threads/hyrule-warriors-rupee-address.505724/) and to an extend in [iAroc's website editor](https://github.com/iAroc/iAroc.github.io/blob/master/hyruleWarriors/index.html#L209) it should be a length of 4. 

I have also made it `02B8` so it is a proper length of 4 just to standardize.

Note that I had to make this change for the recording for the issue I made to the main repo. If you want to confirm my video from that issue with this game please grab my changed script.